smoketest for cal subscription
